airspeed: Generalize interface, multi-instance support
