+++ src/drivers/ets_airspeed/ets_airspeed.cpp
@@ -191,9 +191,9 @@ ETSAirspeed::collect()
 	report.temperature = -1000.0f;
 	report.max_differential_pressure_pa = _max_differential_pressure_pa;
 
-	if (_airspeed_pub > 0 && !(_pub_blocked)) {
+	if (_pub > 0 && !(_pub_blocked)) {
 		/* publish it */
-		orb_publish(ORB_ID(differential_pressure), _airspeed_pub, &report);
+		orb_publish(_orb_id, _pub, &report);
 	}
 
 	new_report(report);
